requirements.txt
numpy
smplx
scipy
pytorch3d
tqdm
matplotlib
aitviewer